Refactored Item(base).jsx to be more felxible
