+++ UI/components/Item.jsx
@@ -1,9 +1,5 @@
 import * as yup from "yup";
 import { Formik, Form, Field } from "formik";
-import { axiosAPI } from "../config/axios";
-import { useRouter } from "next/router";
-import { useContext } from "react";
-import { AlgoliaContext } from "../pages/_app";
 
 /**
  * @typedef Item
@@ -15,7 +11,7 @@ import { AlgoliaContext } from "../pages/_app";
 
 /**
  *
- * @param {{...Item, editable: boolean}} arg0
+ * @param {{...Item, editable: boolean, deletable: boolean, handleDelete: (uuid: string) => void, handleSubmit: ({ values, setSubmitting: (isSubmitting: boolean)=>void }) => }} arg0
  */
 export default function Item({
     uuid,
@@ -23,10 +19,10 @@ export default function Item({
     description,
     sku,
     editable,
+    deletable,
+    handleDelete = () => {},
+    handleSubmit = () => {},
 }) {
-    const router = useRouter();
-    const { searchClient } = useContext(AlgoliaContext);
-
     const itemSchema = yup.object().shape({
         name: yup.string().required("Name is required"),
         description: yup.string().required("Description is required"),
@@ -34,10 +30,6 @@ export default function Item({
     });
 
     if (editable) {
-        const handleDelete = async () => {
-            await axiosAPI.delete(`/items/${uuid}`);
-            router.push("/items");
-        };
         return (
             <>
                 <Formik
@@ -47,13 +39,7 @@ export default function Item({
                         sku,
                     }}
                     validationSchema={itemSchema}
-                    onSubmit={(values, { setSubmitting }) => {
-                        setSubmitting(true);
-                        axiosAPI.put(`/items/${uuid}`, values).then(() => {
-                            setSubmitting(false);
-                            router.push("/items");
-                        });
-                    }}
+                    onSubmit={handleSubmit}
                 >
                     {({ errors, touched, isValid, isSubmitting, dirty }) => (
                         <Form>
@@ -151,13 +137,15 @@ export default function Item({
                                             >
                                                 Save
                                             </button>
-                                            <button
-                                                className="bg-red-500 hover:bg-red-700 text-white font-bold py-2 px-4 rounded ml-4 focus:outline-none focus:shadow-outline"
-                                                type="button"
-                                                onClick={handleDelete}
-                                            >
-                                                Delete
-                                            </button>
+                                            {deletable && (
+                                                <button
+                                                    className="bg-red-500 hover:bg-red-700 text-white font-bold py-2 px-4 rounded ml-4 focus:outline-none focus:shadow-outline"
+                                                    type="button"
+                                                    onClick={handleDelete}
+                                                >
+                                                    Delete
+                                                </button>
+                                            )}
                                         </div>
                                     </div>
                                 </div>
